user authentication for mutation

+++ prisma-node/src/resolvers/Mutation.mjs
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId.mjs'
 
 export const Mutation = { 
   async createUser(parent, args, { prisma }, info) {
@@ -36,16 +37,20 @@ export const Mutation = {
         token: jwt.sign({userId: user.id}, 'thisisasecret')
       }
     },
-    async deleteUser(parent, args, { prisma }, info) {
-      return prisma.deleteUser({id: args.id }, info)
+    async deleteUser(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
+      return prisma.deleteUser({id: userId }, info)
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
       return prisma.updateUser({
-          where: { id: args.id },
+          where: { id: userId },
           data: args.data
         }, info)
     },
-    async createPost(parent, args, { prisma }, info) {
+    async createPost(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
+
       return prisma.createPost(
         {
           title: args.data.title,
@@ -53,28 +58,53 @@ export const Mutation = {
           published: args.data.published,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
         }
       }, info)
     },
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
+      const postExists = await prisma.$exists.post({
+        id: args.id,
+        author: {
+          id: userId
+        }
+      })
+
+      if (!postExists) {
+        throw new Error('Unable to delete post')
+      }
       return prisma.deletePost({id: args.id }, info)
     },
-    updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
+
+      const postExists = await prisma.$exists.post({
+        id: args.id,
+        author: {
+          id: userId
+        }
+      })
+
+      if (!postExists) {
+        throw new Error('Unable to delete post')
+      }
       return prisma.updatePost({
         where: { id: args.id },
         data: args.data
       }, info)
     },
 
-    createComment(parent, args, { prisma }, info) {
+    createComment(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
+
       return prisma.createComment(
         {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           },
           post: {
@@ -84,10 +114,32 @@ export const Mutation = {
           }
       }, info)
     },
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
+      const commentExists = await prisma.$exists.comment({
+        id: args.id,
+        author: {
+          id: userId
+        }
+      })
+
+      if (!commentExists) {
+        throw new Error('Unable to delete comment')
+      }
       return prisma.deleteComment({id: args.id }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, req }, info) {
+      const userId = getUserId(req)
+      const commentExists = await prisma.$exists.comment({
+        id: args.id,
+        author: {
+          id: userId
+        }
+      })
+
+      if (!commentExists) {
+        throw new Error('Unable to delete comment')
+      }
       return prisma.updateComment({
         where: { id: args.id },
         data: args.data
